Refactor - move User model into separate file, move routes into separate file. Add app and views folders.  Include test login,signup pages in views folder
