+++ models/user.js
@@ -1,8 +1,18 @@
+var Sequelize = require('sequelize');
+var sequelize = ()
+
 "use strict";
 
-module.exports = function(sequelize, DataTypes){
+// module.exports = function(sequelize, DataTypes){
+// 	var User = sequelize.define("User", {
+// 		username: sequelize.STRING,
+// 		password: sequelize.STRING,
+// 	});
+// }
+
 	var User = sequelize.define("User", {
 		username: sequelize.STRING,
 		password: sequelize.STRING,
 	});
-}
+
+module.exports = User;
