+++ package.json
@@ -11,6 +11,7 @@
     "method-override": "~2.0.2",
     "mongoose": "~3.8.0",
     "oauth": "^0.9.14",
+    "morgan": "^1.7.0",
     "passport": "^0.3.2",
     "passport-local": "^1.0.0",
     "pg": "^4.4.6",
